requirements.txt
anyio==3.6.2
black==22.10.0
click==8.1.3
ecdsa==0.18.0
fastapi==0.87.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.2
platformdirs==2.5.4
pyasn1==0.4.8
pydantic==1.10.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.19.0
